(refactor:) Refactor kvs -> kvs.manifest, kvs.memory
